double I* revolution.

+++ include/XCM/XBotXDDP.h
@@ -62,49 +62,49 @@ public:
     void update();
     
     // NOTE IXBotJoint getters
-    virtual bool get_link_pos(int joint_id, float& link_pos) final;
+    virtual bool get_link_pos(int joint_id, double& link_pos) final;
     
-    virtual bool get_motor_pos(int joint_id, float& motor_pos) final;
+    virtual bool get_motor_pos(int joint_id, double& motor_pos) final;
     
-    virtual bool get_link_vel(int joint_id, int16_t& link_vel) final;
+    virtual bool get_link_vel(int joint_id, double& link_vel) final;
     
-    virtual bool get_motor_vel(int joint_id, int16_t& motor_vel) final;
+    virtual bool get_motor_vel(int joint_id, double& motor_vel) final;
     
-    virtual bool get_torque(int joint_id, float& torque) final;
+    virtual bool get_torque(int joint_id, double& torque) final;
     
-    virtual bool get_temperature(int joint_id, uint16_t& temperature) final;
+    virtual bool get_temperature(int joint_id, double& temperature) final;
     
-    virtual bool get_fault(int joint_id, uint16_t& fault) final;
+    virtual bool get_fault(int joint_id, double& fault) final;
     
-    virtual bool get_rtt(int joint_id, uint16_t& rtt) final;
+    virtual bool get_rtt(int joint_id, double& rtt) final;
     
-    virtual bool get_op_idx_ack(int joint_id, uint16_t& op_idx_ack) final;
+    virtual bool get_op_idx_ack(int joint_id, double& op_idx_ack) final;
     
-    virtual bool get_aux(int joint_id, float& aux) final;
+    virtual bool get_aux(int joint_id, double& aux) final;
     
-    virtual bool get_gains(int joint_id, std::vector< uint16_t >& gain_vector) final;
+    virtual bool get_gains(int joint_id, std::vector< double >& gain_vector) final;
     
     // NOTE IXBotJoint setters
-    virtual bool set_pos_ref(int joint_id, const float& pos_ref) final;
+    virtual bool set_pos_ref(int joint_id, const double& pos_ref) final;
     
-    virtual bool set_vel_ref(int joint_id, const int16_t& vel_ref) final;
+    virtual bool set_vel_ref(int joint_id, const double& vel_ref) final;
     
-    virtual bool set_tor_ref(int joint_id, const int16_t& tor_ref) final;
+    virtual bool set_tor_ref(int joint_id, const double& tor_ref) final;
     
-    virtual bool set_gains(int joint_id, const std::vector<uint16_t>& gains) final;
+    virtual bool set_gains(int joint_id, const std::vector<double>& gains) final;
     
-    virtual bool set_fault_ack(int joint_id, const int16_t& fault_ack) final;
+    virtual bool set_fault_ack(int joint_id, const double& fault_ack) final;
     
-    virtual bool set_ts(int joint_id, const uint16_t& ts) final;
+    virtual bool set_ts(int joint_id, const double& ts) final;
     
-    virtual bool set_op_idx_aux(int joint_id, const uint16_t& op_idx_aux) final;
+    virtual bool set_op_idx_aux(int joint_id, const double& op_idx_aux) final;
     
-    virtual bool set_aux(int joint_id, const float& aux) final;
+    virtual bool set_aux(int joint_id, const double& aux) final;
     
     // NOTE IXBotFT getters
-    virtual bool get_ft(int ft_id, std::vector< float >& ft, int channels = 6) final;
-    virtual bool get_ft_fault(int ft_id, uint16_t& fault) final;
-    virtual bool get_ft_rtt(int ft_id, uint16_t& rtt) final;
+    virtual bool get_ft(int ft_id, std::vector< double >& ft, int channels = 6) final;
+    virtual bool get_ft_fault(int ft_id, double& fault) final;
+    virtual bool get_ft_rtt(int ft_id, double& rtt) final;
     
     
     // TBD do an interface that does it
